Update tag_hostname.py
An additional if/else error control has been added just to check in case r.json() contains nothing due to an error. In that situation you can't check its status_code and trying would mean a non controlled error stopping the program. With this new if/else all is controlled and the program has a better resilience.